+ update: se actualizo vista usuario con sessiones y panel de admin en home
