+++ adminPanel/templatetags/custom_filters.py
@@ -6,3 +6,7 @@ register = template.Library()
 def formato_moneda(valor):
     valor = int(valor)
     return '${:,.0f}'.format(valor).replace(',', '.')
+
+@register.filter
+def to_int(value):
+    return int(value)
